Python Web Application using Flask Framework/requirements.txt
alembic==1.4.2
certifi==2020.4.5.1
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.2
MarkupSafe==1.1.1
numpy==1.18.1
python-dateutil==2.8.1
python-editor==1.0.4
six==1.14.0
SQLAlchemy==1.3.16
Werkzeug==1.0.1
WTForms==2.3.1
gunicorn==19.0.0
mysql-connector-python==8.0.20
protobuf==3.6.0
